♻️(api-presentation) factorize common error serializers
